fix(tri): Fix build errors from Zig 0.15 module changes (#415)
- Fixed hippocampus stub: added Entry.id(), Entry.summary() methods
- Added writeRule, writeObservation, writeHeartbeat, writeError stubs
- Fixed train_types: step/ppl/loss now use u32/f32 for type compatibility
- Fixed train_diagnostics: added scanCheckpoints() function
- Fixed tri_experiment: changed best_ppl/best_loss/latest_loss to f32
- Fixed phoenix_core: stubbed hippocampus.read() calls in sleepCycle
- Fixed module imports in src/vsa/*.zig to use module system
- Added Sacred.PHI, PHI_INV, PHI_SQ, PHI_INV_SQ, LOG2_3 constants
- Updated metabolism.zig to use correct Sacred field names

Progress: 100+ errors → 3 remaining errors

+++ src/tri/tri_experiment.zig
@@ -88,8 +88,8 @@ fn runChart(args: []const []const u8) void {
     }
 
     // Find global min/max PPL and step range
-    var min_ppl: f32 = 999999.0;
-    var max_ppl: f32 = 0.0;
+    var min_ppl: f64 = 999999.0;
+    var max_ppl: f64 = 0.0;
     var max_step: u32 = 0;
 
     for (0..n_dirs) |i| {
